Add files via upload
List of files produced on the thumbdrive (without a password). These documents reportedly constituted a duplicate copy of the Second FOIA Response from the U.S. Department of Justice in respect of the FOIA Request seeking four categories of records of the speeches of former U.S. Attorney Preet Bharara.

However, there is a notable difference in files between the thumbdrive and the CD-ROM (which was reported to be identical).